Refactor CatImage component to use reusable Card.Image with loading and error states
