+++ src/components/CatImage.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Card } from './ui';
 import type { CatImage as CatImageType } from '../types';
 
@@ -21,28 +21,8 @@ const CatImage: React.FC<CatImageProps> = ({
   aspectRatio = 'square',
   showBreedName = true,
 }) => {
-  const [isLoading, setIsLoading] = useState(true);
-  const [hasError, setHasError] = useState(false);
-
   // Get the first breed if available
   const breed = image.breeds && image.breeds.length > 0 ? image.breeds[0] : null;
-
-  // Handle image load
-  const handleImageLoad = () => {
-    setIsLoading(false);
-  };
-
-  // Handle image error
-  const handleImageError = () => {
-    setIsLoading(false);
-    setHasError(true);
-  };
-
-  // Aspect ratio classes
-  const aspectRatioClasses = {
-    square: 'aspect-square',
-    auto: 'aspect-auto',
-  };
 
   return (
     <Card
@@ -50,47 +30,11 @@ const CatImage: React.FC<CatImageProps> = ({
       hoverable={!!onClick}
       onClick={onClick}
     >
-      <div className={`image-container-aspect ${aspectRatioClasses[aspectRatio]}`}>
-        {/* Loading state */}
-        {isLoading && (
-          <div className="image-loading-container">
-            <div className="image-loading-spinner"></div>
-          </div>
-        )}
-
-        {/* Error state */}
-        {hasError && (
-          <div className="image-loading-container">
-            <div className="text-center">
-              <svg
-                className="image-error-icon"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-                aria-hidden="true"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M6 18L18 6M6 6l12 12"
-                />
-              </svg>
-              <p className="image-error-text">Failed to load image</p>
-            </div>
-          </div>
-        )}
-
-        {/* Image */}
-        <img
-          src={image.url}
-          alt={breed ? `Cat of breed ${breed.name}` : 'Cat image'}
-          className={`image-responsive ${isLoading || hasError ? 'opacity-0' : 'opacity-100'}`}
-          onLoad={handleImageLoad}
-          onError={handleImageError}
-          loading="lazy"
-        />
-      </div>
+      <Card.Image
+        src={image.url}
+        alt={breed ? `Cat of breed ${breed.name}` : 'Cat image'}
+        aspectRatio={aspectRatio === 'square' ? 'square' : 'auto'}
+      />
 
       {/* Breed name if available and requested */}
       {showBreedName && breed && (
